+++ meta/runtime.yml
@@ -9,6 +9,9 @@ action_groups:
   proxmox:
     - proxmox
     - proxmox_access_acl
+    - proxmox_acme_account
+    - proxmox_acme_account_info
+    - proxmox_acme_accounts_info
     - proxmox_backup
     - proxmox_backup_info
     - proxmox_backup_schedule
@@ -17,9 +20,6 @@ action_groups:
     - proxmox_ceph_mon
     - proxmox_ceph_osd
     - proxmox_cluster
-    - proxmox_cluster_acme_account
-    - proxmox_cluster_acme_account_info
-    - proxmox_cluster_acme_accounts_info
     - proxmox_cluster_firewall
     - proxmox_cluster_ha_resources
     - proxmox_cluster_ha_rules
