+++ tests/unit/plugins/modules/test_proxmox_acme_accounts_info.py
@@ -16,18 +16,7 @@ from ansible_collections.community.internal_test_tools.tests.unit.plugins.module
     set_module_args,
 )
 
-from ansible_collections.community.proxmox.plugins.modules import proxmox_cluster_acme_account_info
-
-SAMPLE_GET = {
-    "account": {
-        "contact": ["mailto:example@example.com"],
-        "createdAt": "2024-01-01T00:00:00Z",
-        "status": "valid",
-    },
-    "directory": "https://acme-v02.api.letsencrypt.org/directory",
-    "location": "https://acme-v02.api.letsencrypt.org/acme/acct/3198781481",
-    "tos": "https://letsencrypt.org/documents/LE-SA-v1.3-September-21-2022.pdf",
-}
+from ansible_collections.community.proxmox.plugins.modules import proxmox_acme_accounts_info
 
 
 def exit_json(*args, **kwargs):
@@ -45,15 +34,14 @@ def build_args(**overrides):
         "api_host": "host",
         "api_user": "user",
         "api_password": "password",
-        "name": "default",
         **overrides,
     }
 
 
-class TestProxmoxClusterAcmeAccountInfo(ModuleTestCase):
+class TestProxmoxClusterAcmeAccountsInfo(ModuleTestCase):
     def setUp(self):
         super().setUp()
-        self.module = proxmox_cluster_acme_account_info
+        self.module = proxmox_acme_accounts_info
         self.mock_module_helper = patch.multiple(
             basic.AnsibleModule,
             exit_json=exit_json,
@@ -65,8 +53,7 @@ class TestProxmoxClusterAcmeAccountInfo(ModuleTestCase):
         ).start()
         mock_api = self.connect_mock.return_value
         account_attr = mock_api.cluster.return_value.acme.return_value.account
-        self.named_account = account_attr.return_value
-        self.named_account.return_value = self.named_account
+        self.account_index = account_attr.return_value
 
     def tearDown(self):
         self.connect_mock.stop()
@@ -78,20 +65,20 @@ class TestProxmoxClusterAcmeAccountInfo(ModuleTestCase):
             self.module.main()
         return exc_info.value.args[0]
 
-    def test_get_one(self):
-        self.named_account.get.return_value = SAMPLE_GET
+    def test_list_names(self):
+        self.account_index.get.return_value = [
+            {"name": "staging"},
+            {"name": "default"},
+        ]
 
-        result = self._run_module(build_args(name="default"))
+        result = self._run_module(build_args())
 
         assert result["changed"] is False
-        assert result["name"] == "default"
-        assert result["directory"] == SAMPLE_GET["directory"]
-        assert result["account"]["contact"] == ["example@example.com"]
+        assert result["accounts"] == ["default", "staging"]
 
-    def test_get_one_missing(self):
-        self.named_account.get.side_effect = Exception("404 Not Found: does not exist")
+    def test_list_empty(self):
+        self.account_index.get.return_value = []
 
-        result = self._run_module(build_args(name="missing"))
+        result = self._run_module(build_args())
 
-        assert result.get("failed") is True
-        assert "does not exist" in result["msg"]
+        assert result["accounts"] == []
